Added a Readable Sign
Created a Readable sign prefab - can display a default message typed into the Default message field in the ReadableSign script component attached to the ReadingRange child object.
Imported sign sprite, and signs shadow sprite
Created Message Holder script for random messages. Script is attached to the game manager, and if the sign is set to display random messages, grabs a message from the Gamemanager on start.
Created readable sign script that fades in or out text if player is in the reading range.
Also fixed the shop tilemap collider to combine the wall colliders together to prevent projectiles from going through them.